Long类型比较相等问题调整

+++ src/main/java/com/ruoyi/project/system/user/controller/UserController.java
@@ -181,7 +181,7 @@ public class UserController extends BaseController
         userService.checkUserAllowed(user);
         if (userService.resetUserPwd(user) > 0)
         {
-            if (ShiroUtils.getUserId() == user.getUserId())
+            if (ShiroUtils.getUserId().longValue() == user.getUserId().longValue())
             {
                 setSysUser(userService.selectUserById(user.getUserId()));
             }
